refactor: Extract prefetch function to file
