+++ packages/insight-viewer/src/hooks/useMultiframeImages/types.ts
@@ -1,0 +1,6 @@
+import { CornerstoneImage } from '../../utils/cornerstoneHelper'
+
+export interface Prefetched {
+  image: CornerstoneImage
+  loaded: number
+}
